// src/types/searchCriterias.ts
export type DurationRangeType = [string | null | undefined, string | null | undefined]
export type LabelObject = {
  id: string
  label: string
}