// src/types/requestCriterias.ts
export enum RessourceType {
  REQUEST = 'Request',
  IPP_LIST = 'IPPList',
  PATIENT = 'Patient',
  ENCOUNTER = 'Encounter',
  DOCUMENTS = 'DocumentReference',
  PMSI = 'pmsi',
  CONDITION = 'Condition',
  PROCEDURE = 'Procedure',
  CLAIM = 'Claim',
  MEDICATION = 'Medication',
  MEDICATION_REQUEST = 'MedicationRequest',
  MEDICATION_ADMINISTRATION = 'MedicationAdministration',
  BIO_MICRO = 'biologie_microbiologie',
  OBSERVATION = 'Observation',
  MICROBIOLOGIE = 'microbiologie',
  PHYSIOLOGIE = 'physiologie',
  IMAGING = 'ImagingStudy'
}

export enum RessourceTypeLabels {
  REQUEST = 'Mes requềtes',
  IPP_LIST = "Liste d'IPP",
  PATIENT = 'Démographie',
  ENCOUNTER = 'Prise en charge',
  DOCUMENTS = 'Documents cliniques',
  PMSI = 'PMSI',
  CONDITION = 'Diagnostic',
  PROCEDURE = 'actes CCAM',
  CLAIM = 'GHM',
  MEDICATION = 'Médicament',
  BIO_MICRO = 'Biologie/Microbiologie',
  OBSERVATION = 'Biologie',
  MICROBIOLOGIE = 'Microbiologie',
  PHYSIOLOGIE = 'Physiologie',
  IMAGING = 'Imagerie'
}

export type CommonCriteriaDataType = {
  id: number
  error?: boolean
  type: RessourceType
  encounterService?: ScopeTreeRow[]
  isInclusive?: boolean
  title: string
}

export type SelectedCriteriaType =
  | CcamDataType
  | Cim10DataType
  | DemographicDataType
  | GhmDataType
  | EncounterDataType
  | DocumentDataType
  | MedicationDataType
  | ObservationDataType
  | IPPListDataType
  | ImagingDataType

export type DraftSelectedCriteriaType = SelectedCriteriaType & {
  id?: number
}

export enum CriteriaDataKey {
  GENDER = 'gender',
  VITALSTATUS = 'status',
  PRISE_EN_CHARGE_TYPE = 'priseEnChargeType',
  TYPE_DE_SEJOUR = 'typeDeSejour',
  FILE_STATUS = 'fileStatus',
  ADMISSION_MODE = 'admissionModes',
  ADMISSION = 'admission',
  ENTRY_MODES = 'entryModes',
  EXIT_MODES = 'exitModes',
  REASON = 'reason',
  DESTINATION = 'destination',
  PROVENANCE = 'provenance',
  CIM_10_DIAGNOSTIC = 'cim10Diagnostic',
  DIAGNOSTIC_TYPES = 'diagnosticTypes',
  CCAM_DATA = 'ccamData',
  GHM_DATA = 'ghmData',
  MEDICATION_DATA = 'medicationData',
  PRESCRIPTION_TYPES = 'prescriptionTypes',
  ADMINISTRATIONS = 'administrations',
  BIOLOGY_DATA = 'biologyData',
  MODALITIES = 'modalities',
  DOC_TYPES = 'docTypes',
  STATUS_DIAGNOSTIC = 'statusDiagnostic'
}
// ...
export enum Comparators {
  LESS_OR_EQUAL = '<=',
  LESS = '<',
  EQUAL = '=',
  GREATER = '>',
  GREATER_OR_EQUAL = '>=',
  BETWEEN = '<x>'
}

export type EncounterDataType = CommonCriteriaDataType & {
  type: RessourceType.ENCOUNTER
  age: DurationRangeType
  duration: DurationRangeType
  admissionMode: LabelObject[] | null
  entryMode: LabelObject[] | null
  exitMode: LabelObject[] | null
  priseEnChargeType: LabelObject[] | null
  typeDeSejour: LabelObject[] | null
  fileStatus: LabelObject[] | null
  reason: LabelObject[] | null
  destination: LabelObject[] | null
  provenance: LabelObject[] | null
  admission: LabelObject[] | null
  encounterService: ScopeTreeRow[] | null
  encounterStartDate?: string | null
  encounterEndDate?: string | null
  occurrence: number
  occurrenceComparator: Comparators
  startOccurrence?: string
  endOccurrence?: string
}